Fix speed limit offset
Fix speed limit offset calculation. Previously if offset was 2 (10%), the calculation would have been:

2 + ((5*2)/100) = 2.1
Then if sign speed was 45mph, the set speed would be 95.5mph.

By removing the first offset variable and making it 1, the calculation is 1 + ((5*2)/100) = 1.1. Then if sign speed was 45mph, the set speed would be 49.5mph.